+++ logya/util.py
@@ -3,26 +3,28 @@ import re
 
 import yaml
 
-from collections import namedtuple
 from pathlib import Path
 
-# FIXME document breaking change from dir_*
-# FIXME make root, content and templates required
-cwd = Path.cwd()
-Paths = namedtuple('Paths', ['root', 'content', 'templates', 'static', 'public'])
-paths = Paths(
-    root=cwd,
-    content=Path(cwd, 'content'),
-    templates=Path(cwd, 'templates'),
-    static=Path(cwd, 'static'),
-    public=Path(cwd, 'public')
-)
 
-# Load site config
-config = yaml.load(Path(paths.root, 'site.yaml').read_text(), Loader=yaml.FullLoader)
-
-# Characters not to used in URLs
+# Characters not to be used in URLs
 re_forbidden = re.compile(r'[^\.\w]+')
+
+
+def load_settings():
+    """Return a dictionary with global settings."""
+
+    root = Path.cwd()
+    settings = {
+        'paths': {
+            'root': root,
+            'content': Path(root, 'content'),
+            'templates': Path(root, 'templates'),
+            'static': Path(root, 'static'),
+            'public': Path(root, 'public')
+        }
+    }
+    settings.update(yaml.load(Path(root, 'site.yaml').read_text(), Loader=yaml.FullLoader))
+    return settings
 
 
 def slugify(s: str) -> str:
@@ -33,5 +35,3 @@ def slugify(s: str) -> str:
     Different input strings may result in the same output.
     """
     return re.sub(re_forbidden, '-', s.strip()).strip('-')
-
-